requirements.txt
streamlit
transformers
requests
weaviate-client
torch
python-dotenv